Update Question 8 P174.py

+++ P174 /Question 8 P174.py
@@ -2,13 +2,14 @@
 #Date 2nd September 2025
 #Programme Programme that asks for your height in cm and converts it to feet and inches
 
-heightcm = int(input('Please enter your height in cm: '))
+heightcm = float(input('Please enter your height in cm: '))
 
 heightinch = heightcm/2.54
-heightinch = int(heightinch)
+heightinch = float(heightinch)
 
 heightfoot = heightinch//12
-heightfoot = int(heightfoot)
+heightfoot = float(heightfoot)
 remainder = heightinch%12
 
 print(heightcm,'cm is',heightfoot,'feet',remainder,'inches')
+
